+++ app/src/main/java/com/soturine/volumeok/application/SoundSnapshotGateway.kt
@@ -1,0 +1,7 @@
+package com.soturine.volumeok.application
+
+import com.soturine.volumeok.domain.SoundSnapshot
+
+interface SoundSnapshotGateway : RingVolumeGateway {
+    fun readSnapshot(): SoundSnapshot
+}